Function that matches variants to evidence items

param sample_calls data frame with variant calls param db database with evidence items ('civic','cgi') param colset character vector with column names to pull out from sample_calls param eitems raw list of evidence items param region_marker logical indication if region biomarkers are to be matched or not

Usage

qc_var_eitems(var_eitems = NULL, marker_type = "codon")

Arguments

var_eitems

variant-evidence items

marker_type

type of biomarker

Details

export Function that makes a quality control check of evidence items assigned to variants